Moved DatasetTableIndexer
