+++ backend/dataall/searchproxy/__init__.py
@@ -1,10 +1,8 @@
 from .connect import connect
-from .indexers import upsert_dataset_tables
 from .search import run_query
 
 __all__ = [
     'connect',
     'run_query',
     'upsert',
-    'upsert_dataset_tables',
 ]
