+++ tests/api/conftest.py
@@ -29,7 +29,10 @@ def patch_es(module_mocker):
     module_mocker.patch('dataall.searchproxy.connect', return_value={})
     module_mocker.patch('dataall.searchproxy.search', return_value={})
     module_mocker.patch('dataall.searchproxy.upsert', return_value={})
-    module_mocker.patch('dataall.searchproxy.indexers.upsert_dataset_tables', return_value={})
+    module_mocker.patch(
+        'dataall.modules.datasets.indexers.table_indexer.DatasetTableIndexer.indexers.upsert_all',
+        return_value={}
+    )
     module_mocker.patch('dataall.modules.datasets.indexers.dataset_indexer.DatasetIndexer.upsert', return_value={})
     module_mocker.patch('dataall.searchproxy.indexers.DatasetTableIndexer.upsert', return_value={})
     module_mocker.patch(
